Bug 1502337: Initial support for docker-worker payloads
This commit implements the initial support for docker-worker payloads.
So far, only the Image, Command, Env and MaxRunTime fields are supported.
It hasn't been extensive tested so features like task cancellation may
not work. Also, it doesn't implement custom artifacts and more extensive
tests.

The support for docker-worker payloads is achieved through build flags.
To build with docker-worker paylod, build it with the "docker" flag,
like so:

$ go build -tags=docker

Notice that generic-worker will not be supported in this build.

We also needed to upgrade golang due to a bug in the go compiler [1].

[1] https://github.com/golang/go/issues/25908